+++ xchainer/check_backward.h
@@ -10,6 +10,6 @@ namespace xchainer {
 using Arrays = std::vector<Array>;
 
 void CheckBackwardComputation(std::function<Arrays(const Arrays&)> func, const Arrays& inputs, const Arrays& grad_outputs,
-                              const Arrays& eps, float atol = 1e-5, float rtol = 1e-4);
+                              const Arrays& eps, double atol = 1e-5, double rtol = 1e-4);
 
 }  // namespace xchainer
